gmail_labeling: fix CI (mypy/ruff/async/goldilocks) + address review
CI fixes:
- mypy: type the untyped google-api-python-client service as Any (service.py,
  backend.py) instead of Resource, which has no static .users() and triggered
  no-any-return.
- ruff lint aspect: register the new gmail_api package in ruff.toml
  known-first-party so the sandboxed aspect sorts its imports as first-party;
  reformat the affected planner imports + server.py.
- test_server: add @pypi//pytest_asyncio (async tests need the plugin, not just
  anyio).
- cluster validation: set goldilocks.fairwinds.com/enabled on the gmail-labeling
  namespace.

Review feedback:
- gmail_api/labels.py: use ConfigDict(alias_generator=to_camel) instead of
  hand-written per-field aliases on GmailLabel/CreateLabelRequest.
- Deduplicate the static-bearer ASGI guard: extract mcp_infra/static_bearer.py
  (StaticBearerGuard) and use it from both the oauth_facade and gmail_labeling
  servers instead of copies.
