+++ cluster/k8s/agents/gmail-labeling/namespace.yaml
@@ -2,6 +2,9 @@ apiVersion: v1
 kind: Namespace
 metadata:
   name: gmail-labeling
+  labels:
+    # Single low-traffic server; VPA recommendations add no value here.
+    goldilocks.fairwinds.com/enabled: "false"
   annotations:
     description: >-
       Holds the gmail-labeling MCP server and its Airlock-provisioned gmail.modify
