+++ gmail_archiver/planners/dbsa.py
@@ -5,8 +5,8 @@ from datetime import UTC, datetime, timedelta
 
 from pydantic import BaseModel
 
-from gmail_archiver.event_classifier import EmailTemplateExtractor
 from gmail_api.labels import SystemLabel
+from gmail_archiver.event_classifier import EmailTemplateExtractor
 from gmail_archiver.inbox import GmailInbox
 from gmail_archiver.models import Email
 from gmail_archiver.plan import LABEL_AUTO_CLEANED, Plan
