+++ mcp_infra/oauth_facade/server.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 import asyncio
 import contextlib
-import hmac
 import logging
 import sys
 from collections.abc import AsyncIterator
@@ -17,37 +16,17 @@ from starlette.applications import Starlette
 from starlette.requests import Request
 from starlette.responses import JSONResponse
 from starlette.routing import Mount, Route
-from starlette.types import ASGIApp, Receive, Scope, Send
+from starlette.types import ASGIApp
 
 from mcp_infra.authentik_auth.auth import build_authentik_auth
 from mcp_infra.oauth_facade.config import FacadeSettings
 from mcp_infra.oauth_facade.proxy import build_proxy_server
 from mcp_infra.oauth_facade.upstream_probe import ProbeState, run_probe_loop
 from mcp_infra.persistence import build_client_storage
+from mcp_infra.static_bearer import StaticBearerGuard
 from mcp_infra.tool_filter import ToolFilterMiddleware
 
 logger = logging.getLogger(__name__)
-
-
-class _StaticBearerGuard:
-    """ASGI guard requiring a fixed `Authorization: Bearer <token>` on the wrapped app.
-
-    Wraps only the MCP mount; /healthz and /readyz are registered as earlier
-    routes, so k8s probes reach them without the token. Used for cluster-internal
-    facades configured with `client_auth.static_bearer`.
-    """
-
-    def __init__(self, app: ASGIApp, *, token: str) -> None:
-        self._app = app
-        self._expected = f"Bearer {token}".encode()
-
-    async def __call__(self, scope: Scope, receive: Receive, send: Send) -> None:
-        if scope["type"] == "http":
-            presented = dict(scope["headers"]).get(b"authorization", b"")
-            if not hmac.compare_digest(presented, self._expected):
-                await JSONResponse({"error": "unauthorized"}, status_code=401)(scope, receive, send)
-                return
-        await self._app(scope, receive, send)
 
 
 def build_server(settings: FacadeSettings, *, auth_provider: Any | None = None) -> tuple[Any, Any]:
@@ -59,7 +38,7 @@ def build_server(settings: FacadeSettings, *, auth_provider: Any | None = None) 
     callback, surfacing as `glide_shared.exceptions.TimeoutError: timed out`.
     """
     if settings.client_auth is not None:
-        # Cluster-internal facade: the static bearer (enforced by _StaticBearerGuard
+        # Cluster-internal facade: the static bearer (enforced by StaticBearerGuard
         # in create_app) replaces the Authentik OAuth gate, so there is no OIDC
         # state to persist.
         server = build_proxy_server(settings)
@@ -124,7 +103,7 @@ def create_app(settings: FacadeSettings, *, auth_provider: Any | None = None) ->
 
     mounted_app: ASGIApp = mcp_app
     if settings.client_auth is not None:
-        mounted_app = _StaticBearerGuard(mcp_app, token=settings.client_auth.static_bearer)
+        mounted_app = StaticBearerGuard(mcp_app, token=settings.client_auth.static_bearer)
     return Starlette(
         routes=[Route("/healthz", healthz), Route("/readyz", readyz), Mount("/", app=mounted_app)], lifespan=lifespan
     )
